Remove hard-coded user to access ankrobes
Previously there was only one user - me. Now the user and passwd are
taken from the currently obligatory Basic auth that clients need to
provide to access the service. This will need to be migrated to JWT or
similar but works for the moment.

Also add an /authget endpoint for auth via Basic auth and HTTP GET. This
is required for requiring Basic auth with the nginx-ingress kube ingress
controller to simulate the current apache WSGI basic auth hack

Also add /hello endpoint. This is needed for the basic kube http probes.

Signed-off-by: Anton Melser <anton@transcrob.es>