Major rewrite, integrate ankrobes functionality
New features:
- Initial support added for English for Chinese speakers
- Integrate ankrobes-server functionality directly in transcrobes (via
        django-wsgi)
- Implement JWT for the enrich endpoints
- Almost complete refactoring to allow for multiple languages
- Removal of many Chinese specific elements (many still to fix, like
        calling the phonetic representation 'pinyin' still)
- Allow for keeping dictionaries and other data sources in the DB rather
than in memory (inmem a config option). This allows for having multiple
transcrobes instances with multiple languages. Performance seems to drop
by about 10-15% but instances are MUCH more stable and kube can manage
them like it should because now they are 'stateless'.
- A loader django command 'load_data' was written for one-shot loading
or reloading of the static data files to the DB. It is possible to force
reloading if the import mechanism has changed (so changing what would be
        stored in the DB)
- Many tests were written before the rewrite but with so many major
changes, most now need significant modification. That is definititely
finally almost certainly probably next. :-)

Signed-off-by: Anton Melser <anton@transcrob.es>