Add fix for stale refresh token
When first getting a page from cache older than the exp of the refresh
token, the rxdb will be inited with an invalid refresh token. Subsequent
proxy.init calls will be made (on page navigations) that have up-to-date
refresh tokens, so we can simply replace. Also have refresh token calls
return a new refresh token to extend their life as much as possible.

There are still cases where this isn't ideal but as everything will be
migrated to an SPA, it definitely doesn't merit major thought at this
time (because it should be as secure as what we have now).

Signed-off-by: Anton Melser <anton@transcrob.es>