Fix bug wiping defs in local map of existing words
In order to have a list of existing words, when a word was practiced a
definition was being taken from the list of all unseen/new defs. But
revisions weren't in that list, and the items were already in the list,
resulting in existing words being overwritten with undefined. That is
very dumb

Signed-off-by: Anton Melser <anton@transcrob.es>